First-class custom elements, cell child improvements
First, the restriction on cell children (eg. if a cell is added as a
child of another element) that demanded they be cells containing vectors
or seqs of DOM nodes has been removed. It is now possible to do things
like this:

    (defc thing "asdf")

    (div
      (h1 thing)
      (p "This is interesting."))

See below for some good applications of this new ability.

This commit introduces first-class custom elements. First-classness in
this case means fully supporting the DOM element semantic for the full
lifecycle of the custom element.

For example, suppose foo is a custom element. You can now do:

    (foo
      :class "some-class"
      :title (cell= (if (even? num-cell) "bar" "baz"))
      (loop-tpl
        :bindings [thing things]
        (span :text thing)))

or,

    ((foo "one" "two") :title "heyo" "three" "four")

This was previously not possible because the children of the custom
element needed to be provided explicitly at creation time. In this case
we don't have all the elements at creation time; elements will be added
and removed dynamically throughout the lifetime of the element.

The new, first-class custom elements are created by the #'component
macro (for anonymous custom elements) or the corresponding defining
form, #'defcomponent.

    (defcomponent foo
      [{:keys [title]} [intro & kids]]
      (div
        (h1 (text title))
        (p intro)
        (ul
          (loop-tpl
            :bindings [kid kids]
            (li kid)))))

What's different about this is that the arguments are bound to cells,
not values. These cells will change over time when elem.appendChild(),
elem.removeChild(), elem.setAttribute() etc. are called. This is why
loop-tpl is used to populate the children of the ul; they will be added
and removed dynamically, and kids is a cell.

To clarify: the title, intro, and kids bindings in the example above
are, in fact, cells, not values.