Don't ever call jQuery#remove() on elements
The user should do this, instead. When Hoplon does it automatically it
can interfere with the operation of jQuery plugins that move elements
into and out of the DOM (because when Hoplon calls .remove() on the
element all event handlers are removed, which can break the plugin).

The main use case was to accomodate using cell= in place of loop-tpl,
like this:

        (ul (cell= (map li items)))

This can still be accomplished with a little scaffolding:

        (ul (doto (cell= (map li items))
              (add-watch (gensym)
                (fn [_ _ prev _]
                  (.remove (js/jQuery prev))))))

Or use loop-tpl and avoind needing to call .remove() at all:

        (ul (loop-tpl :bindings [i items] (li i)))